# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Cenos API
  description: |
    API do ecossistema Cenos para gestão de agentes de campo (leituristas)
    das concessionárias de energia do Piauí (PI) e Maranhão (MA).

    ## Autenticação

    A API suporta 5 métodos de autenticação:

    1. **JWT Bearer** — Painel administrativo (`Authorization: Bearer <token>`)
    2. **Telegram TMA** — App do agente (`X-Telegram-Init-Data`)
    3. **Token Simples** — Consultas externas/BI (`?token=API_TOKEN`)

    Consulte a documentação detalhada em `docs/` para cada módulo.
  version: "2.0.0"
  contact:
    name: Cenos Suporte
    url: https://cenos.com.br
servers:
  - url: https://api.izi.tec.br
    description: Produção
  - url: http://localhost:3040
    description: Desenvolvimento
...
paths:
...
  /admin/security_reports:
    get:
      tags: [Admin]
      summary: Lista relatórios de segurança
      security:
        - JwtAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: estado
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Relatórios paginados
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SecurityReport"
                  total:
                    type: integer
                  page:
                    type: integer
                  totalPages:
                    type: integer
    post:
      tags: [Admin]
      summary: Cria relatório de segurança
      security:
        - JwtAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                autor:
                  type: string
                motivo:
                  type: string
                observacao:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                estado:
                  type: string
      responses:
        "201":
          description: Relatório criado
  /admin/security_reports/{id}:
    delete:
      tags: [Admin]
      summary: Remove relatório de segurança
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Relatório removido

  # ─── ADMIN: BADGES ───────────────────────────────────────────────────
...
  /admin/messages/notifications/{agentId}:
    get:
      tags: [Admin, Notificações]
      summary: Histórico de notificações do agente
      security:
        - JwtAuth: []
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Notificações do agente

  # ─── ADMIN: CHAT ─────────────────────────────────────────────────────
...
  /admin/permission:
    get:
      tags: [Admin]
      summary: Lista perfis de permissão
      security:
        - JwtAuth: []
      responses:
        "200":
          description: Perfis
    post:
      tags: [Admin]
      summary: Cria perfil de permissão
      security:
        - JwtAuth: []
      responses:
        "201":
          description: Perfil criado

  # ─── ADMIN: MESSAGE TEMPLATES ────────────────────────────────────────
...
  /agent/security_report:
    post:
      tags: [Agente]
      summary: Enviar relatório de segurança
      security:
        - TelegramAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                motivo:
                  type: string
                  example: "Assalto na região"
                observacao:
                  type: string
                  example: "Presença de suspeitos na rota"
                latitude:
                  type: number
                  example: -5.08921
                longitude:
                  type: number
                  example: -42.80167
      responses:
        "201":
          description: Relatório de segurança registrado
    get:
      tags: [Agente]
      summary: Listar relatórios de segurança do agente
      security:
        - TelegramAuth: []
      responses:
        "200":
          description: Relatórios de segurança
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

  # ─── AGENTE: UPLOAD ──────────────────────────────────────────────────
...
  /admin/user/users/{id}/permissions:
    put:
      tags: [Admin]
      summary: Atualizar permissões do usuário
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions:
                  type: array
                  items:
                    type: string
                  example: ["dashboard", "users", "config"]
      responses:
        "200":
          description: Permissões atualizadas
        "404":
          description: Usuário não encontrado

  # ─── ADMIN: NOTIFICATIONS ────────────────────────────────────────────
  /admin/notifications/send:
    post:
      tags: [Admin, Notificações]
      summary: Enviar notificação para agente específico
      security:
        - JwtAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_id:
                  type: string
                  example: "12345"
                title:
                  type: string
                  example: "Nova meta"
                body:
                  type: string
                  example: "Você foi designado para nova rota"
                type:
                  type: string
                  enum: [success, warn, danger, info]
      responses:
        "200":
          description: Notificação enviada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  id:
                    type: integer
  /admin/notifications/broadcast:
    post:
      tags: [Admin, Notificações]
      summary: Enviar notificação para todos agentes
      security:
        - JwtAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Aviso geral"
                body:
                  type: string
                  example: "Manutenção programada domingo"
                type:
                  type: string
                  enum: [success, warn, danger, info]
                estado:
                  type: string
                  enum: [pi, ma]
      responses:
        "200":
          description: Broadcast enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  total:
                    type: integer
                    example: 150

  # ─── ADMIN: APP PINS ────────────────────────────────────────────────
...
  /admin/permission/{id}:
    get:
      tags: [Admin]
      summary: Detalhes de perfil de permissão
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Perfil de permissão
        "404":
          description: Não encontrado
    delete:
      tags: [Admin]
      summary: Remover perfil de permissão
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Perfil removido
        "404":
          description: Não encontrado
...
components:
...
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticação administrativa via JWT.
        Header: `Authorization: Bearer <token>`
        O token é obtido via `POST /admin/user/login`.
    TelegramAuth:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description: |
        Autenticação do agente via Telegram Mini App.
        Header: `X-Telegram-Init-Data: <initData>`
        O initData é validado via Telegram TMA.
    ApiTokenAuth:
      type: apiKey
      in: query
      name: token
      description: |
        Token simples para consultas externas.
        Query param: `?token=API_TOKEN`
  schemas:
...
    SecurityReport:
      type: object
      properties:
        id:
          type: integer
        autor:
          type: string
        motivo:
          type: string
        observacao:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        estado:
          type: string
        created_at:
          type: string
          format: date-time
        nome:
          type: string
        matricula:
          type: string
...